Add link object controls to Match view
